backend/requirements.txt
Flask==3.1.2
Flask-Cors==6.0.1
python-dotenv==1.2.1
requests==2.32.5
google-generativeai==0.8.5
